Siirrä VktSuoritus ja YkiSuoritus tiedonsiirtoschema-pakettiin ja sinetöi KielitutkinnonSuoritus
Sinetöity (sealed) rajapinta vaatii Kotlinissa, että alityypit ovat
samassa paketissa. Tämän jälkeen Henkilosuoritus.toEntity ja
ValidationService.validateAndEnrich saavat tyhjentävän when-haaran ja
fallback-virhehaarat voidaan poistaa.
