+++ server/src/main/kotlin/fi/oph/kitu/tiedonsiirtoschema/KielitutkinnonSuoritus.kt
@@ -4,8 +4,7 @@ import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
 import fi.oph.kitu.koodisto.Koodisto
 import fi.oph.kitu.oid.Oid
-import fi.oph.kitu.vkt.VktSuoritus
-import fi.oph.kitu.yki.suoritukset.YkiSuoritus
+import fi.oph.kitu.vkt.VktSuoritusEntity
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDate
 import java.time.OffsetDateTime
@@ -17,17 +16,6 @@ data class Henkilosuoritus<T : KielitutkinnonSuoritus>(
     val lisatty: OffsetDateTime? = null,
 ) {
     fun modifySuoritus(f: (T) -> T): Henkilosuoritus<T> = copy(suoritus = f(suoritus))
-
-    inline fun <reified A> toEntity(): A =
-        when (suoritus) {
-            is VktSuoritus -> suoritus.toVktSuoritusEntity(henkilo)
-
-            is YkiSuoritus -> suoritus.toYkiSuoritusEntity(henkilo)
-
-            else -> throw RuntimeException(
-                "Failed to convert to entity: not implemented for Henkilosuoritus<${suoritus::class.simpleName}>",
-            )
-        } as A
 }
 
 @JsonTypeInfo(
@@ -44,7 +32,7 @@ interface PolymorphicByTyyppi {
     JsonSubTypes.Type(value = VktSuoritus::class, name = "valtionhallinnonkielitutkinto"),
     JsonSubTypes.Type(value = YkiSuoritus::class, name = "yleinenkielitutkinto"),
 )
-interface KielitutkinnonSuoritus :
+sealed interface KielitutkinnonSuoritus :
     PolymorphicByTyyppi,
     Lahdejarjestelmallinen {
     override val tyyppi: Koodisto.SuorituksenTyyppi
