+++ server/src/main/kotlin/fi/oph/kitu/util/validation/ValidationService.kt
@@ -1,16 +1,15 @@
 package fi.oph.kitu.util.validation
 
 import arrow.core.NonEmptyList
-import arrow.core.nonEmptyListOf
 import arrow.core.raise.either
 import fi.oph.kitu.tiedonsiirtoschema.Henkilosuoritus
 import fi.oph.kitu.tiedonsiirtoschema.HenkilosuoritusValidation
 import fi.oph.kitu.tiedonsiirtoschema.KielitutkinnonSuoritus
-import fi.oph.kitu.vkt.VktSuoritus
+import fi.oph.kitu.tiedonsiirtoschema.VktSuoritus
+import fi.oph.kitu.tiedonsiirtoschema.YkiSuoritus
 import fi.oph.kitu.vkt.VktValidation
 import fi.oph.kitu.yki.arvioijat.YkiArvioija
 import fi.oph.kitu.yki.arvioijat.YkiArvioijaValidation
-import fi.oph.kitu.yki.suoritukset.YkiSuoritus
 import fi.oph.kitu.yki.suoritukset.YkiSuoritusValidation
 import org.springframework.stereotype.Service
 
@@ -21,36 +20,25 @@ final class ValidationService(
     val ykiSuoritus: YkiSuoritusValidation,
     val ykiArvioija: YkiArvioijaValidation,
 ) {
-    inline fun <reified T : KielitutkinnonSuoritus> validateAndEnrich(
-        hs: Henkilosuoritus<T>,
-    ): ValidationResult<Henkilosuoritus<T>> =
+    fun <T : KielitutkinnonSuoritus> validateAndEnrich(hs: Henkilosuoritus<T>): ValidationResult<Henkilosuoritus<T>> =
         either {
-            with(commonValidation) { validateAndEnrich(hs) }
+            val common = with(commonValidation) { validateAndEnrich(hs) }
+            val enriched =
+                when (val s = common.suoritus) {
+                    is VktSuoritus -> {
+                        with(vkt) {
+                            validateAndEnrich(Henkilosuoritus(common.henkilo, s, common.lisatty))
+                        }
+                    }
+
+                    is YkiSuoritus -> {
+                        with(ykiSuoritus) {
+                            validateAndEnrich(Henkilosuoritus(common.henkilo, s, common.lisatty))
+                        }
+                    }
+                }
             @Suppress("UNCHECKED_CAST")
-            when (hs.suoritus) {
-                is VktSuoritus -> {
-                    with(vkt) {
-                        validateAndEnrich(Henkilosuoritus(hs.henkilo, hs.suoritus))
-                    } as Henkilosuoritus<T>
-                }
-
-                is YkiSuoritus -> {
-                    with(ykiSuoritus) {
-                        validateAndEnrich(Henkilosuoritus(hs.henkilo, hs.suoritus))
-                    } as Henkilosuoritus<T>
-                }
-
-                else -> {
-                    raise(
-                        nonEmptyListOf(
-                            Validation.ValidationError(
-                                emptyList(),
-                                "Validation not implemented for ${hs::class.simpleName}",
-                            ),
-                        ),
-                    )
-                }
-            }
+            enriched as Henkilosuoritus<T>
         }
 
     fun validateAndEnrich(arvioija: YkiArvioija): ValidationResult<YkiArvioija> =
