+++ server/src/main/kotlin/fi/oph/kitu/vkt/html/VktErinomaisenArviointiPage.kt
@@ -20,8 +20,8 @@ import fi.oph.kitu.i18n.finnishDateTimeUTC
 import fi.oph.kitu.koodisto.Koodisto
 import fi.oph.kitu.oppijanumero.OppijanumeroException
 import fi.oph.kitu.oppijanumero.OppijanumerorekisteriHenkilo
+import fi.oph.kitu.tiedonsiirtoschema.VktHenkilosuoritus
 import fi.oph.kitu.util.growToSize
-import fi.oph.kitu.vkt.VktHenkilosuoritus
 import fi.oph.kitu.vkt.VktOsakoe
 import kotlinx.html.FlowContent
 import kotlinx.html.footer
