--- server/src/main/kotlin/fi/oph/kitu/yki/suoritukset/YkiSuoritus.kt
@@ -1,117 +1,0 @@
-package fi.oph.kitu.yki.suoritukset
-
-import fi.oph.kitu.koodisto.Koodisto
-import fi.oph.kitu.oid.Oid
-import fi.oph.kitu.tiedonsiirtoschema.Henkilo
-import fi.oph.kitu.tiedonsiirtoschema.Henkilosuoritus
-import fi.oph.kitu.tiedonsiirtoschema.KielitutkinnonSuoritus
-import fi.oph.kitu.tiedonsiirtoschema.LahdejarjestelmanTunniste
-import fi.oph.kitu.yki.Arviointitila
-import fi.oph.kitu.yki.TutkinnonOsa
-import fi.oph.kitu.yki.Tutkintokieli
-import fi.oph.kitu.yki.Tutkintotaso
-import java.lang.IllegalArgumentException
-import java.time.Instant
-import java.time.LocalDate
-
-typealias YkiHenkilosuoritus = Henkilosuoritus<YkiSuoritus>
-
-data class YkiSuoritus(
-    val tutkintotaso: Tutkintotaso,
-    val kieli: Tutkintokieli,
-    val todistuskieli: Todistuskieli?,
-    val jarjestaja: YkiJarjestaja,
-    val tutkintopaiva: LocalDate,
-    val arviointipaiva: LocalDate?,
-    val osat: List<YkiOsa>,
-    val tarkistusarviointi: YkiTarkastusarviointi? = null,
-    val arviointitila: Arviointitila,
-    override val lahdejarjestelmanId: LahdejarjestelmanTunniste,
-    override val internalId: Int? = null,
-    override val koskiOpiskeluoikeusOid: Oid? = null,
-    override val koskiSiirtoKasitelty: Boolean = false,
-) : KielitutkinnonSuoritus {
-    override val tyyppi: Koodisto.SuorituksenTyyppi = Koodisto.SuorituksenTyyppi.YleinenKielitutkinto
-
-    fun toYkiSuoritusEntity(henkilo: Henkilo): YkiSuoritusEntity {
-        fun arvosana(tyyppi: TutkinnonOsa): Int? =
-            osat
-                .firstOrNull { it.tyyppi == tyyppi }
-                ?.arvosana
-
-        return YkiSuoritusEntity(
-            id = internalId,
-            suorittajanOID = henkilo.oid,
-            hetu = henkilo.hetu,
-            sukupuoli = henkilo.sukupuoli ?: throw IllegalArgumentException("Sukupuoli puuttuu"),
-            sukunimi = henkilo.sukunimi ?: throw IllegalArgumentException("Sukunimi puuttuu"),
-            etunimet = henkilo.etunimet ?: throw IllegalArgumentException("Etunimet puuttuu"),
-            kansalaisuus = henkilo.kansalaisuus ?: throw IllegalArgumentException("Kansalaisuus puuttuu"),
-            katuosoite = henkilo.katuosoite ?: throw IllegalArgumentException("Katuosoite puuttuu"),
-            postinumero = henkilo.postinumero ?: throw IllegalArgumentException("Postinumero puuttuu"),
-            postitoimipaikka = henkilo.postitoimipaikka ?: throw IllegalArgumentException("Postitoimipaikka puuttuu"),
-            maa = henkilo.maa,
-            email = henkilo.email,
-            solkiId = lahdejarjestelmanId.id.toInt(),
-            lastModified = Instant.now(),
-            tutkintopaiva = tutkintopaiva,
-            tutkintokieli = kieli,
-            tutkintotaso = tutkintotaso,
-            todistuskieli = todistuskieli,
-            jarjestajanTunnusOid = jarjestaja.oid,
-            jarjestajanNimi = jarjestaja.nimi,
-            arviointipaiva = arviointipaiva,
-            tekstinYmmartaminen = arvosana(TutkinnonOsa.tekstinYmmartaminen),
-            kirjoittaminen = arvosana(TutkinnonOsa.kirjoittaminen),
-            rakenteetJaSanasto = arvosana(TutkinnonOsa.rakenteetJaSanasto),
-            puheenYmmartaminen = arvosana(TutkinnonOsa.puheenYmmartaminen),
-            puhuminen = arvosana(TutkinnonOsa.puhuminen),
-            yleisarvosana = arvosana(TutkinnonOsa.yleisarvosana),
-            tarkistusarvioinninSaapumisPvm = tarkistusarviointi?.saapumispaiva,
-            tarkistusarvioinninAsiatunnus = tarkistusarviointi?.asiatunnus,
-            tarkistusarvioidutOsakokeet = tarkistusarviointi?.tarkistusarvioidutOsakokeet?.toSet(),
-            arvosanaMuuttui = tarkistusarviointi?.arvosanaMuuttui?.toSet(),
-            perustelu = tarkistusarviointi?.perustelu,
-            tarkistusarvioinninKasittelyPvm = tarkistusarviointi?.kasittelypaiva,
-            tarkistusarviointiHyvaksyttyPvm = null,
-            koskiOpiskeluoikeus = koskiOpiskeluoikeusOid,
-            koskiSiirtoKasitelty = koskiSiirtoKasitelty,
-            arviointitila = arviointitila,
-            arviointitilaLahetetty = null,
-            arviointitilanLahetysvirhe = null,
-        )
-    }
-}
-
-data class YkiJarjestaja(
-    val oid: Oid,
-    val nimi: String,
-)
-
-data class YkiOsa(
-    val tyyppi: TutkinnonOsa,
-    val arvosana: Int?,
-)
-
-data class YkiTarkastusarviointi(
-    val saapumispaiva: LocalDate,
-    val kasittelypaiva: LocalDate?,
-    val asiatunnus: String,
-    val tarkistusarvioidutOsakokeet: List<TutkinnonOsa>?,
-    val arvosanaMuuttui: List<TutkinnonOsa>?,
-    val perustelu: String,
-) {
-    companion object {
-        fun from(s: YkiSuoritusEntity): YkiTarkastusarviointi? =
-            s.tarkistusarvioinninSaapumisPvm?.let {
-                YkiTarkastusarviointi(
-                    it,
-                    s.tarkistusarvioinninKasittelyPvm,
-                    s.tarkistusarvioinninAsiatunnus.orEmpty(),
-                    s.tarkistusarvioidutOsakokeet?.toList(),
-                    s.arvosanaMuuttui?.toList(),
-                    s.perustelu.orEmpty(),
-                )
-            }
-    }
-}
